refac: mv list_files and validate_source_target_files to data_module_utils
