+++ src/careamics/compat/dataset/dataset_utils/__init__.py
@@ -4,14 +4,10 @@ __all__ = [
     "create_write_file_path",
     "get_files_size",
     "iterate_over_files",
-    "list_files",
-    "validate_source_target_files",
 ]
 
 from .file_utils import (
     create_write_file_path,
     get_files_size,
-    list_files,
-    validate_source_target_files,
 )
 from .iterate_over_files import iterate_over_files
