+++ src/careamics/compat/lightning/train_data_module.py
@@ -11,11 +11,7 @@ from torch.utils.data import DataLoader, IterableDataset
 
 from careamics.compat.config.augmentations import TransformConfig
 from careamics.compat.config.data import DataConfig
-from careamics.compat.dataset.dataset_utils import (
-    get_files_size,
-    list_files,
-    validate_source_target_files,
-)
+from careamics.compat.dataset.dataset_utils import get_files_size
 from careamics.compat.dataset.in_memory_dataset import (
     InMemoryDataset,
 )
@@ -24,6 +20,10 @@ from careamics.compat.dataset.iterable_dataset import (
 )
 from careamics.config.support import SupportedData
 from careamics.file_io.read import get_read_func
+from careamics.lightning.data_module_utils import (
+    list_files,
+    validate_source_target_files,
+)
 from careamics.utils import get_logger, get_ram_size
 
 DatasetType = Union[InMemoryDataset, PathIterableDataset]
