+++ src/careamics/lightning/data_module_utils.py
@@ -1,6 +1,7 @@
 """Utility functions for file and paths solver."""
 
 from collections.abc import Sequence
+from fnmatch import fnmatch
 from pathlib import Path
 from typing import Any, Literal
 
@@ -8,10 +9,6 @@ from numpy import ndarray
 from numpy.typing import NDArray
 from typing_extensions import TypeIs
 
-from careamics.compat.dataset.dataset_utils.file_utils import (
-    list_files,
-    validate_source_target_files,
-)
 from careamics.config.support import SupportedData as SD
 from careamics.dataset.factory import ImageStackLoading, ReadFuncLoading
 from careamics.dataset.image_stack_loader.zarr_utils import is_valid_uri
@@ -142,6 +139,119 @@ def _is_path_data(
     input_is_valid = _is_path_input(data[0])
     target_is_valid = data[1] is None or _is_path_input(data[1])
     return input_is_valid and target_is_valid
+
+
+def list_files(
+    data_path: str | Path,
+    data_type: str | SD,
+    extension_filter: str = "",
+) -> list[Path]:
+    """List recursively files in `data_path` and return a sorted list.
+
+    If `data_path` is a file, its name is validated against the `data_type` using
+    `fnmatch`, and the method returns `data_path` itself.
+
+    By default, if `data_type` is equal to `custom`, all files will be listed. To
+    further filter the files, use `extension_filter`.
+
+    `extension_filter` must be compatible with `fnmatch` and `Path.rglob`, e.g. "*.npy"
+    or "*.czi".
+
+    Parameters
+    ----------
+    data_path : Union[str, Path]
+        Path to the folder containing the data.
+    data_type : Union[str, SupportedData]
+        One of the supported data type (e.g. tif, custom).
+    extension_filter : str, optional
+        Extension filter, by default "".
+
+    Returns
+    -------
+    list[Path]
+        list of pathlib.Path objects.
+
+    Raises
+    ------
+    FileNotFoundError
+        If the data path does not exist.
+    ValueError
+        If the data path is empty or no files with the extension were found.
+    ValueError
+        If the file does not match the requested extension.
+    """
+    # convert to Path
+    data_path = Path(data_path)
+
+    # raise error if does not exists
+    if not data_path.exists():
+        raise FileNotFoundError(f"Data path {data_path} does not exist.")
+
+    # get extension compatible with fnmatch and rglob search
+    extension = SD.get_extension_pattern(data_type)
+
+    if data_type == SD.CUSTOM and extension_filter != "":
+        extension = extension_filter
+
+    # search recurively
+    if data_path.is_dir() and data_path.suffix != ".zarr":
+        # search recursively the path for files with the extension
+        files = sorted(data_path.rglob(extension))
+    else:
+        # raise error if it has the wrong extension
+        if not fnmatch(str(data_path.absolute()), extension):
+            raise ValueError(
+                f"File {data_path} does not match the requested extension "
+                f'"{extension}".'
+            )
+
+        # save in list
+        files = [data_path]
+
+    # raise error if no files were found
+    if len(files) == 0:
+        raise ValueError(
+            f'Data path {data_path} is empty or files with extension "{extension}" '
+            f"were not found."
+        )
+
+    return files
+
+
+def validate_source_target_files(src_files: list[Path], tar_files: list[Path]) -> None:
+    """
+    Validate source and target path lists.
+
+    The two lists should have the same number of files, and the filenames should match.
+
+    Parameters
+    ----------
+    src_files : list of pathlib.Path
+        List of source files.
+    tar_files : list of pathlib.Path
+        List of target files.
+
+    Raises
+    ------
+    ValueError
+        If the number of files in source and target folders is not the same.
+    ValueError
+        If some filenames in Train and target folders are not the same.
+    """
+    # check equal length
+    if len(src_files) != len(tar_files):
+        raise ValueError(
+            f"The number of source files ({len(src_files)}) is not equal to the number "
+            f"of target files ({len(tar_files)})."
+        )
+
+    # check identical names
+    src_names = {f.name for f in src_files}
+    tar_names = {f.name for f in tar_files}
+    difference = src_names.symmetric_difference(tar_names)
+
+    if len(difference) > 0:
+        raise ValueError(f"Source and target files have different names: {difference}.")
 
 
 def list_files_in_directory(
